move arround the packages
