+++ app/src/main/kotlin/com/wire/android/ui/home/conversations/details/AddMembersToPeopleViewModel.kt
@@ -1,28 +1,24 @@
-package com.wire.android.ui.home.newconversation
+package com.wire.android.ui.home.conversations.details
 
 import androidx.lifecycle.SavedStateHandle
 import com.wire.android.R
 import com.wire.android.mapper.ContactMapper
-import com.wire.android.navigation.EXTRA_CONVERSATION_ID
 import com.wire.android.navigation.NavigationManager
+import com.wire.android.ui.home.newconversation.search.SearchPeopleViewModel
+import com.wire.android.ui.home.newconversation.search.SearchResult
 import com.wire.android.util.dispatchers.DispatcherProvider
 import com.wire.kalium.logic.data.id.ConversationId
-import com.wire.kalium.logic.data.id.QualifiedID
-import com.wire.kalium.logic.data.id.parseIntoQualifiedID
-import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.feature.connection.SendConnectionRequestUseCase
-import com.wire.kalium.logic.feature.conversation.AddMemberToConversationUseCase
 import com.wire.kalium.logic.feature.conversation.GetAllContactsNotInConversationUseCase
 import com.wire.kalium.logic.feature.publicuser.search.SearchKnownUsersUseCase
 import com.wire.kalium.logic.feature.publicuser.search.SearchUsersUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.withContext
 import javax.inject.Inject
 import com.wire.kalium.logic.feature.conversation.Result as GetContactsResult
 import com.wire.kalium.logic.feature.publicuser.search.Result
 
 @HiltViewModel
-class AddMembersToConversationViewModel @Inject constructor(
+class AddMembersToPeopleViewModel @Inject constructor(
     private val getAllContactsNotInConversation: GetAllContactsNotInConversationUseCase,
     private val searchKnownUsers: SearchKnownUsersUseCase,
     private val searchPublicUsers: SearchUsersUseCase,
@@ -32,7 +28,7 @@ class AddMembersToConversationViewModel @Inject constructor(
     sendConnectionRequest: SendConnectionRequestUseCase,
     savedStateHandle: SavedStateHandle,
     navigationManager: NavigationManager
-) : SearchConversationViewModel(navigationManager, sendConnectionRequest, dispatchers) {
+) : SearchPeopleViewModel(navigationManager, sendConnectionRequest, dispatchers) {
 
 //    val conversationId: QualifiedID = savedStateHandle
 //        .get<String>(EXTRA_CONVERSATION_ID)!!
@@ -75,7 +71,7 @@ class AddMembersToConversationViewModel @Inject constructor(
                 SearchResult.Success(result.userSearchResult.result.map { otherUser -> contactMapper.fromOtherUser(otherUser) })
             }
         }
-
+//
 //    suspend fun addSelectedMembers() {
 //        withContext(dispatchers.io()) {
 //            addMemberToConversation(conversationId, state.contactsAddedToGroup.map { UserId(it.id, it.domain) })
